Move device info to a entity base class
